refactor(test-loop): move peer_manager_actor.rs from utils to setup
- peer_manager_actor.rs contains setup infrastructure (TestLoopPeerManagerActor,
  TestLoopNetworkSharedState, network message handlers) used during test-loop
  node initialization, not reusable utilities
- All primary consumers are in setup/: setup.rs, state.rs, builder.rs, drop_condition.rs
- Move it to test-loop-tests/src/setup/ where it belongs alongside other setup code
- Update module declarations and imports accordingly
