+++ test-loop-tests/src/setup/mod.rs
@@ -1,5 +1,6 @@
 pub mod builder;
 pub mod drop_condition;
 pub mod env;
+pub(crate) mod peer_manager_actor;
 pub mod setup;
 pub mod state;
